refactoring slide data organization
