+++ ozkour-back/google-api/slide.js
@@ -3,6 +3,7 @@ const utilitary = require('../utilitary')
 const connect = require('./connect.js')
 const dayjs = require('dayjs')
 const customParseFormat = require('dayjs/plugin/customParseFormat')
+const slideDataOrganizer = require('./slideDataOrganizer.js')
 dayjs.extend(customParseFormat)
 
 const presentationId = '1Mwzl0-13stcTZRn_0iyIJLZveuY80SW2cmv9p2Wgpug'
@@ -29,16 +30,6 @@ const greyForegroundColor = {
   }
 }
 
-const slideSpacing = {
-  EVENT: 50,
-  TALK: 45,
-  DATE: 40
-}
-
-const DEFAULT_START_Y_INDEX = 100
-
-const END_OF_SLIDE = 520
-
 async function createSlideFromTalks (talks, h) {
   try {
     const res = await connect.authMethode(createSlides, talks)
@@ -46,178 +37,6 @@ async function createSlideFromTalks (talks, h) {
   } catch (e) {
     return h.response(e).code(500)
   }
-}
-
-/**
- * cluster all the talks by date
- * @param {Array} The talks that need to be clustered
- * @return {dataOrganized} dataOrganized where the keys are the date and the values are the talks
- */
-function clusterByDate (data) {
-  const dataOrganized = new Map()
-  for (let i = 0; i < data.length; i++) {
-    if (!dataOrganized.has(data[i].date)) {
-      dataOrganized.set(data[i].date, [
-        {
-          universe: data[i].universe,
-          eventType: data[i].eventType,
-          eventName: data[i].eventName,
-          talkTitle: data[i].talkTitle,
-          speakers: data[i].speakers
-        }
-      ])
-    } else {
-      const newValue = dataOrganized.get(data[i].date)
-      newValue.push({
-        universe: data[i].universe,
-        eventType: data[i].eventType,
-        eventName: data[i].eventName,
-        talkTitle: data[i].talkTitle,
-        speakers: data[i].speakers
-      })
-    }
-  }
-  return divideInMultipleSlides(clusterByEventName(dataOrganized))
-}
-
-function divideInMultipleSlides (dataOrganized) {
-  const mapIter = dataOrganized.keys()
-  let date = mapIter.next().value
-  let isEndOfData = false
-  let isEndOfSlideReached = false
-  let yNextElmt = DEFAULT_START_Y_INDEX
-  let yNextElmtTemp
-  let DoesDateFits
-  const dataOrganizedBySlides = []
-  dataOrganizedBySlides.push(new Map())
-
-  while (!isEndOfData) {
-    isEndOfSlideReached = false
-    while (!isEndOfData && !isEndOfSlideReached) {
-      DoesDateFits = true
-      yNextElmtTemp = yNextElmt
-      const resDate = tryDate(yNextElmt, dataOrganized.get(date))
-      DoesDateFits = resDate.DoesDateFits
-      yNextElmt = resDate.yNextElmt
-
-      if (!DoesDateFits) { // toute la date ne rentre pas
-        yNextElmt = yNextElmtTemp
-        let i = 0
-        let listOfEventThatFits = []
-        const allEventsInADate = dataOrganized.get(date)
-        let atLeastOneEventCanFit = false
-        let worthToContinue = (i === 0 || atLeastOneEventCanFit)
-        let DoesEventFits = true
-
-        while (i < allEventsInADate.length && worthToContinue) {
-          worthToContinue = (i === 0 || atLeastOneEventCanFit)
-          const event = allEventsInADate[i]
-          const resEvent = tryEvent(yNextElmt, event)
-          DoesEventFits = resEvent.DoesEventFits
-          yNextElmt = resEvent.yNextElmt
-          if (DoesEventFits) {
-            atLeastOneEventCanFit = true
-            listOfEventThatFits.push(event)
-          } else {
-            if (atLeastOneEventCanFit) {
-              // on ajoute ce qui rentre
-              dataOrganizedBySlides[dataOrganizedBySlides.length - 1].set(date, listOfEventThatFits)
-
-              // on prepare la prochaine fois que ça rentre
-              listOfEventThatFits = []
-              const map = new Map()
-              dataOrganizedBySlides.push(map)
-              listOfEventThatFits.push(event)
-              yNextElmt = DEFAULT_START_Y_INDEX
-            } else {
-              i = allEventsInADate.length
-            }
-          }
-          i++
-          if (i >= allEventsInADate.length && atLeastOneEventCanFit) {
-            dataOrganizedBySlides[dataOrganizedBySlides.length - 1].set(date, listOfEventThatFits)
-          }
-        }
-        if (!atLeastOneEventCanFit) {
-          isEndOfSlideReached = true
-          const map = new Map()
-          map.set(date, dataOrganized.get(date))
-          dataOrganizedBySlides.push(map)
-          yNextElmt = DEFAULT_START_Y_INDEX
-        }
-      } else { // toute la date rentre
-        dataOrganizedBySlides[dataOrganizedBySlides.length - 1].set(date, dataOrganized.get(date))
-      }
-
-      // on verifie s'il reste des données
-      date = mapIter.next().value
-      if (date === undefined) {
-        isEndOfData = true
-      }
-    }
-  }
-  return dataOrganizedBySlides
-}
-
-function tryDate (yNextElmt, data) {
-  yNextElmt += slideSpacing.DATE
-  data.forEach(event => {
-    yNextElmt = tryEvent(yNextElmt, event).yNextElmt
-  })
-  return { yNextElmt, DoesDateFits: yNextElmt <= END_OF_SLIDE }
-}
-
-function tryEvent (yNextElmt, event) {
-  yNextElmt += slideSpacing.EVENT
-  event.talks.forEach(talk => {
-    yNextElmt += slideSpacing.TALK
-  })
-
-  return { yNextElmt, DoesEventFits: yNextElmt <= END_OF_SLIDE }
-}
-
-/**
- * cluster all the talks by name of events
- * @param {dataOrganized} The talks already clustured by dates
- * @return {dataOrganized} dataOrganized where the keys are the date and the values are an object
- * with 2 attributes: one for the event name and one for the left data
- */
-function clusterByEventName (dataOrganized) {
-  const mapIter = dataOrganized.keys()
-  let date = mapIter.next().value
-
-  while (date !== undefined) {
-    const EventNameAdded = []
-    const EventArray = []
-    for (let i = 0; i < dataOrganized.get(date).length; i++) {
-      // pour chaque date
-      const talk = dataOrganized.get(date)[i]
-
-      if (EventNameAdded.includes(talk.eventName)) {
-        EventArray[EventNameAdded.indexOf(talk.eventName)].talks.push({
-          universe: talk.universe,
-          talkTitle: talk.talkTitle,
-          speakers: talk.speakers
-        })
-      } else {
-        EventArray.push({
-          eventName: talk.eventName,
-          eventType: talk.eventType,
-          talks: [
-            {
-              universe: talk.universe,
-              talkTitle: talk.talkTitle,
-              speakers: talk.speakers
-            }
-          ]
-        })
-        EventNameAdded.push(talk.eventName)
-      }
-    }
-    dataOrganized.set(date, EventArray)
-    date = mapIter.next().value
-  }
-  return dataOrganized
 }
 
 async function createSlides (auth, talks) {
@@ -229,7 +48,7 @@ async function createSlides (auth, talks) {
       },
       async (err, res) => {
         if (err) reject(err.message)
-        const dataOrganizedBySlide = clusterByDate(talks)
+        const dataOrganizedBySlide = slideDataOrganizer.clusterByDate(talks)
         dataOrganizedBySlide.forEach(dataOrganized =>
           copySlide(auth, res.data.slides[0].objectId, dataOrganized)
             .then((result) => {
@@ -595,14 +414,14 @@ function addTableData (auth, idPage, dataOrganized) {
 
   let date = mapIter.next().value
   let IndexRowInTableToInsert = 0
-  let yNextElmt = DEFAULT_START_Y_INDEX
+  let yNextElmt = slideDataOrganizer.DEFAULT_START_Y_INDEX
 
   while (date !== undefined) {
     const dateId = date.replaceAll('/', '-') + Math.random().toString(36).replaceAll('.', ':')
     IndexRowInTableToInsert = 0
     const dateFormated = date.substring(0, 2) + ' ' + utilitary.convDateToMonth(date) + ' ' + date.substring(6)
     requests.push(addDateTextWithStyle(idPage, dateFormated, dateId, yNextElmt))
-    yNextElmt += slideSpacing.DATE
+    yNextElmt += slideDataOrganizer.slideSpacing.DATE
     requests.push(
       CreateTableWithStyleForAllEventsInDate(
         idPage,
@@ -625,7 +444,7 @@ function addTableData (auth, idPage, dataOrganized) {
       )
       IndexRowInTableToInsert++
 
-      yNextElmt += slideSpacing.EVENT
+      yNextElmt += slideDataOrganizer.slideSpacing.EVENT
 
       // add all talk for the event
       for (let j = 0; j < arrayOfTalksForAnEvent.talks.length; j++) {
@@ -634,7 +453,7 @@ function addTableData (auth, idPage, dataOrganized) {
           addTalkTitleWithStyleToTable(dateId, talk, IndexRowInTableToInsert),
           addSpeakersWithStyleToTable(dateId, talk, IndexRowInTableToInsert)
         )
-        yNextElmt += slideSpacing.TALK
+        yNextElmt += slideDataOrganizer.slideSpacing.TALK
         IndexRowInTableToInsert++
       }
     }
@@ -773,7 +592,5 @@ function copySlide (auth, idPage, talkSelected) {
 }
 
 module.exports = {
-  createSlideFromTalks,
-  clusterByDate,
-  clusterByEventName
+  createSlideFromTalks
 }
